// src/types/recipe.ts
export interface Recipe {
  idMeal: string;
  strMeal: string;
  strDrinkAlternate?: string;
  strCategory: string;
  strArea: string;
  strInstructions: string;
  strMealThumb: string;
  strTags?: string;
  strYoutube?: string;
  [key: string]: string | undefined;
}

export interface MealResponse {
  meals: Recipe[] | null;
}

export interface MealType {
  name: string;
  value: string;
}